feature: create worker/task for email delivery
