+++ internal/notifications/notifier.go
@@ -1,24 +1,23 @@
 package notifications
 
 import (
-	"time"
+	"encoding/json"
+
+	"github.com/hibiken/asynq"
 
 	"github.com/earaujoassis/space/internal/config"
 	"github.com/earaujoassis/space/internal/logs"
-	"github.com/earaujoassis/space/internal/mailer"
 	"github.com/earaujoassis/space/internal/utils"
+	"github.com/earaujoassis/space/internal/workers"
 )
 
 type Notifier struct {
 	cfg *config.Config
-	ms  *mailer.Mailer
 }
 
 func NewNotifier(cfg *config.Config) *Notifier {
-	ms := mailer.NewMailer(cfg)
 	return &Notifier{
 		cfg: cfg,
-		ms:  ms,
 	}
 }
 
@@ -26,62 +25,24 @@ func NewNotifier(cfg *config.Config) *Notifier {
 //
 //	using e-mail messages (production-only) or stdout (development-only)
 func (n *Notifier) Announce(name string, data utils.H) {
-	if n.cfg.Environment == "production" {
-		var err error
-		switch name {
-		case "user.created":
-			data["Year"] = time.Now().Year()
-			message := mailer.CreateMessage("user.created.html", data)
-			err = n.ms.SendEmail(
-				"Welcome to quatroLABS services",
-				message,
-				data["Email"].(string))
-		case "user.update.password":
-			data["Year"] = time.Now().Year()
-			message := mailer.CreateMessage("user.update.password.html", data)
-			err = n.ms.SendEmail(
-				"A magic link to update your password was requested at quatroLABS",
-				message,
-				data["Email"].(string))
-		case "user.update.secrets":
-			data["Year"] = time.Now().Year()
-			message := mailer.CreateMessage("user.update.secrets.html", data)
-			err = n.ms.SendEmail(
-				"A magic link to recreat your recovery code and secret code generator was requested at quatroLABS",
-				message,
-				data["Email"].(string))
-		case "user.update.email_verification":
-			data["Year"] = time.Now().Year()
-			message := mailer.CreateMessage("user.update.email_verification.html", data)
-			err = n.ms.SendEmail(
-				"Please confirm you e-mail address at quatroLABS",
-				message,
-				data["Email"].(string))
-		case "session.created":
-			data["Year"] = time.Now().Year()
-			message := mailer.CreateMessage("session.created.html", data)
-			err = n.ms.SendEmail(
-				"A new session created at quatroLABS",
-				message,
-				data["Email"].(string))
-		case "session.magic":
-			data["Year"] = time.Now().Year()
-			message := mailer.CreateMessage("session.magic.html", data)
-			err = n.ms.SendEmail(
-				"A magic link for a new session was requested at quatroLABS",
-				message,
-				data["Email"].(string))
-		}
+	switch n.cfg.Environment {
+	case "production":
+		cfg := n.cfg
+		enqueuer := asynq.NewClient(asynq.RedisClientOpt{
+			Addr: cfg.MemoryDNS(),
+			DB:   cfg.MemorystoreIndex,
+		})
+		defer enqueuer.Close()
+		payload, err := json.Marshal(workers.EmailDeliveryPayload{
+			Name: name,
+			Data: data,
+		})
 		if err != nil {
-			logs.Propagatef(
-				logs.Critical,
-				"Critical error sending announcement: action `%s` with data `%v`\n",
-				name,
-				data)
+			logs.Propagatef(logs.Error, "could not enqueue task for email delivery: %s", name)
+			return
 		}
-
-	}
-	if n.cfg.Environment == "development" {
+		enqueuer.Enqueue(asynq.NewTask(workers.TypeEmailDelivery, payload))
+	default:
 		logs.Propagatef(logs.Info, "Action `%s` with data `%v`\n", name, data)
 	}
 }
